refactor to centralize promotion callbacks in a plugin
That worked out more nicely in the end.
